refactor: auction 목록 Item에 대해서 공유 컴포넌트 추가
